+++ features/libraries/ui/src/commonMain/kotlin/app/campfire/libraries/ui/detail/composables/DurationListItem.kt
@@ -6,7 +6,9 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -18,10 +20,13 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import app.campfire.common.compose.extensions.clockFormat
 import app.campfire.common.compose.extensions.thenIf
+import app.campfire.common.compose.icons.CampfireIcons
+import app.campfire.common.compose.icons.rounded.Warning
 import app.campfire.core.extensions.fluentIf
 import kotlin.time.Duration
 
@@ -36,6 +41,7 @@ internal fun DurationListItem(
   title: String,
   duration: Duration,
   modifier: Modifier = Modifier,
+  isValid: Boolean = true,
   progress: Float = 0f,
   progressColor: Color = MaterialTheme.colorScheme.secondary,
   trackColor: Color = MaterialTheme.colorScheme.surface.copy(alpha = 0.50f),
@@ -88,10 +94,21 @@ internal fun DurationListItem(
       },
     verticalAlignment = Alignment.CenterVertically,
   ) {
+    if (!isValid) {
+      Icon(
+        CampfireIcons.Rounded.Warning,
+        contentDescription = "Invalid Chapter",
+        modifier = Modifier.size(18.dp),
+        tint = MaterialTheme.colorScheme.error,
+      )
+      Spacer(Modifier.width(8.dp))
+    }
     Text(
       text = title,
       style = MaterialTheme.typography.labelLarge,
-      fontWeight = if (isActiveChapter) FontWeight.Bold else null,
+      fontWeight = if (isActiveChapter || !isValid) FontWeight.Bold else null,
+      fontStyle = if (!isValid) FontStyle.Italic else null,
+      color = if (!isValid) MaterialTheme.colorScheme.error else Color.Unspecified,
       modifier = Modifier.weight(1f),
     )
     Spacer(Modifier.width(16.dp))