+++ features/libraries/ui/src/commonTest/kotlin/app/campfire/libraries/ui/detail/BaseLibraryItemPresenterTest.kt
@@ -12,6 +12,7 @@ import app.campfire.core.model.preview.libraryItem
 import app.campfire.core.session.UserSession
 import app.campfire.libraries.api.screen.LibraryItemScreen
 import app.campfire.libraries.test.FakeLibraryItemRepository
+import app.campfire.libraries.test.FakeLibraryItemValidator
 import app.campfire.playlists.api.dialog.AddToPlaylistDialog
 import app.campfire.series.test.FakeSeriesRepository
 import app.campfire.sessions.test.FakeSessionQueue
@@ -29,6 +30,7 @@ abstract class BaseLibraryItemPresenterTest {
   internal val screen = LibraryItemScreen(TestLibraryItemId)
   internal val navigator = FakeNavigator(screen)
   internal val libraryItemRepository = FakeLibraryItemRepository()
+  internal val libraryItemValidator = FakeLibraryItemValidator()
   internal val seriesRepository = FakeSeriesRepository()
   internal val sessionsRepository = FakeSessionsRepository()
   internal val sessionQueue = FakeSessionQueue()
@@ -48,6 +50,7 @@ abstract class BaseLibraryItemPresenterTest {
     screen = screen,
     navigator = navigator,
     repository = libraryItemRepository,
+    validator = libraryItemValidator,
     seriesRepository = seriesRepository,
     sessionsRepository = sessionsRepository,
     sessionQueue = sessionQueue,