+++ features/sessions/ui/src/commonMain/kotlin/app/campfire/sessions/ui/expanded/ExpandedPlaybackUiState.kt
@@ -3,11 +3,13 @@ package app.campfire.sessions.ui.expanded
 import androidx.compose.runtime.Stable
 import app.campfire.core.model.LibraryItem
 import app.campfire.core.model.LibraryItemId
+import app.campfire.libraries.api.LibraryItemValidation
 import com.slack.circuit.runtime.CircuitUiEvent
 import com.slack.circuit.runtime.CircuitUiState
 
 @Stable
 data class ExpandedPlaybackUiState(
+  val validation: LibraryItemValidation,
   val queue: List<LibraryItem>,
   val reorderSink: suspend (fromItemId: LibraryItemId, toItemId: LibraryItemId) -> Unit,
   val eventSink: (ExpandedPlaybackUiEvent) -> Unit,