+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Revamped the sleep timer bottom sheet UI in the player view
+- Warnings when an item has misaligned chapter information that needs to be corrected on the server
 
 ### Deprecated
 
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Series on the item detail page not appearing until loading them in the series screen
 - App crashing on system reboot due to offline download resumption
 - Mark Finished / Discard Progress would pause playback of current item
+- Playback issues when chapter information is mis-aligned with the audio tracks
 
 ### Other Notes & Contributions